compile done, but media pad has not initialized

+++ drv/fpga_vir_cam.h
@@ -8,8 +8,8 @@
 #define FPGA_VIR_CAM_V4L2_REG_WRITE 0x08000001//(V4L2_CID_PRIVATE_BASE + 1)
 
 struct fpga_vir_cam_reg {
-    unsigned short reg_value;
-    unsigned int reg_addr;
+    unsigned int reg_value;
+    unsigned short reg_addr;
 };
